css-text-3_dev/xhtml1/overflow-wrap-break-word-surrogate-pairs-001.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8" />
<title>CSS Text Test: overflow-wrap: break-word with fit-content</title>
<link href="kojiishi@gmail.com" rel="author" title="Koji Ishii" />
<link href="http://www.w3.org/TR/css-text-3/#overflow-wrap" rel="help" title="6.2. Overflow Wrapping: the 'word-wrap'/'overflow-wrap' property" />
<meta content="dom" name="flags" />
<meta content="The 'overflow-wrap: break-word' must not break surrogate pairs" name="assert" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
.fit {
  display: inline-block;
  word-wrap: break-word;
  font-size: 20px;
  line-height: 1;
}
</style>
</head><body>
<p classs="instructions">Tests pass if all lines do not wrap.
</p><ol id="container">
  <li><div class="fit">ğœ‹</div>
  </li><li><div class="fit">ï¿½ï¿½</div>
</li></ol>
<script>
var tests = container.children;
for (var i = 0; i &lt; tests.length; i++) {
  var element = tests[i];
  test(function () {
    var rects = element.getClientRects();
    console.log(rects[0].width);
    assert_equals(1, rects.length);
    assert_equals(rects[0].height, 20);
  }, (i + 1).toString());
}
</script>

</body></html>
